Workaround softrend 4D render function calls
This is a simple pointer cast.
The incoming 4thed dimension (outv+3) bleeds into a different parameter.
This is usually an indicator that something very wrong is happening.
In this case it uses a parameter declared as "br_uint_16 *fp_vertices".
I can not really make heads or tails about this but I also don't
understand what a four dimensional render is supposed to look like.