Fix compilation with GCC/Clang and LTO and dynamic linking under linux + Packaging (#10)
* Remove not existing headers from cmake file list

Removed reference to:
* core/inc/brlog.h
* core/pixelmap/ddi/memloops.h

* Workaround softrend 4D render function calls

This is a simple pointer cast.
The incoming 4thed dimension (outv+3) bleeds into a different parameter.
This is usually an indicator that something very wrong is happening.
In this case it uses a parameter declared as "br_uint_16 *fp_vertices".
I can not really make heads or tails about this but I also don't
understand what a four dimensional render is supposed to look like.

* Fix softrend v1model array pointers parameters

Casts `br_uint_16 (*)[3]` arrays into `br_uint_16 *` pointers for LTO

* Fix device_pixelmaps not indendified as pixelmaps

`br_device_pixelmap_glf *` and `br_device_pixelmap_virtualfb *` have to be
cast as `br_pixelmap *` as C does not have inheritence.
This is important for LTO besides being generally being good practice.

* Fix fpsetup and pfpsetup doubles

Internally doubles are handled as fp64_t unions.

* Fix softrend temp_face not forward declared

Not forward declaring `struct temp_face` causes temp_face to be
added as new unrelated symbol into each translation unit.

* Fix GCC throwing error at const promotion

For some reason GCC does not like handing a `GLChar **` into a `const
GLchar * const *` parameter. My guess is that the double pointer causes
issues. Anyway, this commit makes GCC happy and it doesn't hurt.
This specific type is what the OpenGL API wants.

* Fix LTO symbol mismatch due to variadic declaration

TriangleRender_ZT_I8_D16_256 is declared as a variadic function.
This adds nothing specific but causing LTO to fail.

* Fix variable br_diaghandler override with GCC

GCC and Clang do not allow for overwritting symbols by link order.
To support this with GCC and Clang the project has to be either
compiled with -fcommon (which would be global, also I have not tested this)
or the overrideable symbol has to be flagged as weak.

* Port various soft render primitives from ASM to C

This commit ports the MASM code for:
* TrapezoidRenderPITAN
* TrapezoidRenderPITA
* TrapezoidRenderPITIA
* TrapezoidRenderPITA15
* TrapezoidRenderPITIA_RGB_555
* TrapezoidRenderPITA24
* TrapezoidRenderPITIA_RGB_888
* TrapezoidRenderPIZ2TAN
* TrapezoidRenderPIZ2TA
* TrapezoidRenderPIZ2TIA
* TrapezoidRenderPIZ2TA15
* TrapezoidRenderPIZ2TIA_RGB_555
* TrapezoidRenderPIZ2TA24
* TrapezoidRenderPIZ2TIA_RGB_888
* ScanLinePITPD64
* ScanLinePITPD128
* ScanLinePITPD256
* ScanLinePITPD1024
* ScanLinePITIP256_RGB_555
* ScanLinePITIPB256_RGB_555
* ScanLinePITIPB256_RGB_565

* Fix cmake include installation directory

* Add packaging

* Replace some types with brender types

* Fix CMake targets and variables

* Variables are now BRender_ instead of BRENDER
* Added CPack support
* Properly export x86emu target

* Update BRenderConfig.cmake.in

Co-authored-by: Anonymous Maarten <madebr@users.noreply.github.com>

* Fix MSVC void* pointer size compile error

* Update cmake build install

---------

Co-authored-by: Jeff Harris <jeff@1amstudios.com>
Co-authored-by: Anonymous Maarten <madebr@users.noreply.github.com>